PIM-5600: Moving ProductQueryUtility class to Pim\Bundle\CatalogBundle
