+++ src/Pim/Bundle/CatalogBundle/Doctrine/MongoDBODM/QueryGenerator/AttributeDeletedQueryGenerator.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\QueryGenerator;
 
-use Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\ProductQueryUtility;
+use Pim\Bundle\CatalogBundle\ProductQueryUtility;
 
 /**
  * Attribute deleted query generator
