+++ src/Pim/Bundle/CatalogBundle/Doctrine/MongoDBODM/QueryGenerator/ChannelDeletedQueryGenerator.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\QueryGenerator;
 
-use Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\ProductQueryUtility;
+use Pim\Bundle\CatalogBundle\ProductQueryUtility;
 
 /**
  * Channel deleted query generator
