+++ src/Pim/Bundle/CatalogBundle/ProductQueryUtility.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM;
+namespace Pim\Bundle\CatalogBundle;
 
 use Pim\Component\Catalog\Model\AttributeInterface;
 use Pim\Component\Catalog\Model\ProductValueInterface;
