+++ src/Pim/Bundle/ReferenceDataBundle/DataGrid/Datasource/ResultRecord/MongoDbOdm/Product/ReferenceDataTransformer.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\ReferenceDataBundle\DataGrid\Datasource\ResultRecord\MongoDbOdm\Product;
 
-use Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\ProductQueryUtility;
+use Pim\Bundle\CatalogBundle\ProductQueryUtility;
 use Pim\Component\Catalog\AttributeTypes;
 
 /**
