+++ src/Pim/Bundle/ReferenceDataBundle/Doctrine/MongoDB/Filter/ReferenceDataFilter.php
@@ -3,7 +3,7 @@
 namespace Pim\Bundle\ReferenceDataBundle\Doctrine\MongoDB\Filter;
 
 use Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\Filter\AbstractAttributeFilter;
-use Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\ProductQueryUtility;
+use Pim\Bundle\CatalogBundle\ProductQueryUtility;
 use Pim\Bundle\ReferenceDataBundle\Doctrine\ReferenceDataIdResolver;
 use Pim\Component\Catalog\Exception\InvalidArgumentException;
 use Pim\Component\Catalog\Model\AttributeInterface;
