+++ src/Pim/Bundle/ReferenceDataBundle/Doctrine/MongoDB/Sorter/ReferenceDataSorter.php
@@ -3,7 +3,7 @@
 namespace Pim\Bundle\ReferenceDataBundle\Doctrine\MongoDB\Sorter;
 
 use Doctrine\ODM\MongoDB\Query\Builder as QueryBuilder;
-use Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\ProductQueryUtility;
+use Pim\Bundle\CatalogBundle\ProductQueryUtility;
 use Pim\Component\Catalog\Model\AttributeInterface;
 use Pim\Component\Catalog\Query\Sorter\AttributeSorterInterface;
 use Pim\Component\ReferenceData\ConfigurationRegistryInterface;
