+++ src/Pim/Component/Catalog/Model/AbstractProduct.php
@@ -420,22 +420,6 @@ abstract class AbstractProduct implements ProductInterface
     /**
      * {@inheritdoc}
      */
-    public function getAttributeByCode($attributeCode)
-    {
-        $attribute = null;
-
-        foreach ($this->values as $value) {
-            if ($attributeCode === $value->getAttribute()->getCode()) {
-                $attribute = $value->getAttribute();
-            }
-        }
-
-        return $attribute;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function getValues()
     {
         $values = new ArrayCollection();
